separate statistics collection for Unity project type
